refactor(hush-cli): extract dispatch from pkg_cli
Move the public cmd_pkg dispatcher into pkg_cli/dispatch.rs and re-export
it via `pub use dispatch::cmd_pkg;` to preserve the pkg_cli::cmd_pkg path
(used by main.rs). dispatch.rs reaches the per-subcommand handlers
through an internal `use super::*`; the handler imports remain in mod.rs.
Write import dropped from mod.rs (its only consumer moved).
